+++ .github/workflows/unit-tests.yml
@@ -29,4 +29,4 @@ jobs:
     - uses: actions/checkout@v2
 
     - name: run go unit tests
-      run: go test ./pq/*/ ./pq/*/*/
+      run: go test ./promq/*/ ./promq/*/*/
