+++ promq/autocomplete/earley/completion_context.go
