+++ promq/autocomplete/earley/suggested_types_test.go
